Fix fail to capture on RDMA device on FreeBSD
In pcap for RDMA devices, on FreeBSD where we have support for capsicum
but we don't have file descriptor for a live capture we will fail with:
"tcpdump: unable to limit pcap descriptor"
We just fail to set the proper access rights - don't fail the capture.

Signed-off-by: Slava Shwartsman <slavash@mellanox.com>

Fix fail to capture on RDMA device on FreeBSD

In pcap for RDMA devices, on FreeBSD where we have support for capsicum
but we don't have file descriptor for a live capture we will fail with:
"tcpdump: unable to limit pcap descriptor"
We just fail to set the proper access rights - don't fail the capture.

Signed-off-by: Slava Shwartsman <slavash@mellanox.com>