Rx: Fix remaining fall-through warnings (GH #618).
Benjamin Kaduk explains in the issue comments:

The first two ("bulk stat" and "callback"; lines 1023 and 1282) should
not fall through, but the other two ("list entry" in various forms,
lines 1741 and 1790) should fall through. The answer lies in the RPC-L
files at [1], [2] and [3] (also line 396), looking at the appropriate IN
or OUT arguments.

1: http://git.openafs.org/?p=openafs.git;a=blob;f=src/fsint/afsint.xg;h=9fe4980c6cffcae457789e363393825e29dcb7b1;hb=HEAD#l594
2: http://git.openafs.org/?p=openafs.git;a=blob;f=src/fsint/afscbint.xg;h=19760bc512c647595662c30e190303b3a20be983;hb=HEAD#l22
3: http://git.openafs.org/?p=openafs.git;a=blob;f=src/vlserver/vldbint.xg;h=ff34d12d45454de328439cab3ed896fd564f0c59;hb=HEAD#l341